网络通信 V2.0.0
为了适应服务器的要求，此处进行大改，内容包括：
重新组织文件结构
将网络收发和数据报处理从一个类中分离出来，分成两个类：
NetworkForClient：收发数据，与服务器通信
MessageOperate：数据报处理
