+++ Client/src/Network_Client/NetworkForClient/Log.txt
@@ -21,3 +21,11 @@ V1.1.0
 应要求，协议格式更改为每条信息包括信息类型与内容
 （注意：好友列表的协议格式保持不变）
 重新生成JavaDoc
+
+2018-4-11
+V2.0.0
+为了适应服务器的模式，这边进行修整，包括：
+重新组织文件结构
+将网络收发和数据报处理从一个类中分离出来，分成两个类：
+NetworkForClient：收发数据，与服务器通信
+MessageOperate：数据报处理
